JCF: Hacks in order to get the code to build given the merge of patch/fddaq-v5.3.x
This is a followup commit after my initial merge of patch/fddaq-v5.3.x, needed in order
for the code to build. It involves me updating the signature of TPReplayApplication::generate_modules
thanks to the changes originally made on patch/fddaq-v5.3.x related to DUNE-DAQ/appmodel#201. It
also involves me literally blocking out the existing TPReplayApplication code using preprocessor
directives as I don't have the needed expertise to overhaul the code in light of the interface change.